enhancement - project strucutre refactor
