+++ components/MainAllies.js
