--- components/app/Home.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import homeComponents from '../../utils/home-components';
-
-export default function Home({ components }) {
-  return (
-    <>
-      {components.map(({ fields, sys: { id, contentType: { sys: { id: sectionType } } } }) => {
-        const Component = homeComponents[sectionType];
-        if (!Component) return null;
-        return <Component key={id} fields={fields} />;
-      })}
-    </>
-  );
-}
